Finished Requirement 04

+++ src/main/java/com/example/repository/MessageRepository.java
@@ -1,9 +1,12 @@
 package com.example.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import com.example.entity.Message;
 
-public interface MessageRepository extends JpaRepository<Message, Long>{
-    
+public interface MessageRepository extends JpaRepository<Message, Integer>{
 }
